ingester/requirements.txt
numpy<2.0
fastapi
uvicorn
python-multipart
faiss-cpu==1.7.4
pymupdf
sentence-transformers
